Correct CMakeLists for gcc/nvcc combined compile
